// src/core/saving_study/domain/interfaces/sort_saving_study.ts
import type { Order } from "@/src/core/app/domain/models/order";

export interface SortSavingStudy {
  id?: Order;
  cups?: Order;
  clientType?: Order;
  clientName?: Order;
  clientNif?: Order;
  marketer?: Order;
  selectedRate?: Order;
  rateType?: Order;
  status?: Order;
  createdAt?: Order;
  responsible?: Order;
}